:D fix navigation, combine yaw & roll
