+++ app/(evaluationComponents)/evaluationYaw.jsx
@@ -10,24 +10,24 @@ import React, { useState, useEffect, useRef } from 'react';
 import { Camera } from 'expo-camera';
 import * as FaceDetector from 'expo-face-detector';
 import * as FileSystem from 'expo-file-system';
-import 'expo-router/entry';
 import { Link, useRouter } from 'expo-router';
 
-const evaluationRoll = () => {
+const evaluationYaw = () => {
   const router = useRouter();
 
   const [hasPermission, setHasPermission] = useState(null);
   const [landmarkData, setLandmarkData] = useState([]);
-  const [roll, setRoll] = useState(0);
+  const [yaw, setYaw] = useState(0);
   const [ScreenText, setScreenText] = useState('');
-  const [maxRollR, setMaxRollR] = useState(0);
-  const [maxRollL, setMaxRollL] = useState(0);
+  const [maxYawR, setMaxYawR] = useState(0);
+  const [maxYawL, setMaxYawL] = useState(0);
   const [lineCoordinates, setLineCoordinates] = useState(null); // Store line coordinates
   const cameraRef = useRef(null);
   const [evaluationStarted, setEvaluationStarted] = useState(true);
   const mutex = useRef(false);
-  const [isRollStable, setIsRollStable] = useState(true);
+  const [isYawStable, setIsYawStable] = useState(true);
   const [cacheBuster, setCacheBuster] = useState(Date.now());
+  const [capturedPhotoUri, setCapturedPhotoUri] = useState(null);
 
   useEffect(() => {
     (async () => {
@@ -62,42 +62,41 @@ const evaluationRoll = () => {
     }
   };
 
-  // Run this effect whenever 'Roll' changes. Capture a picture when maxRollR or maxRollL gets a new value
+  // Run this effect whenever 'yaw' changes. Capture a picture when maxYawR or maxYawL gets a new value
   useEffect(() => {
-    if (!isRollStable) {
+    if (!isYawStable) {
       console.log('Warte auf das Erreichen der Endposition...');
       return;
     }
 
     let stableTimer;
 
-    if (roll >= 0 && roll < 80) {
-      console.log(`Current RollR: ${roll}, Max RollR: ${maxRollR}`);
-      setScreenText('-> ' + roll + '°');
-      setMaxRollR((prev) => {
-        if (roll > prev) {
-          console.log('Taking picture for MaxRollR');
-          takePicture('MaxRollR.jpg');
+    if (yaw > 15 && yaw < 180) {
+      console.log(`Current YawR: ${yaw}, Max YawR: ${maxYawR}`);
+      setScreenText('-> ' + yaw + '°');
+      setMaxYawR((prev) => {
+        if (yaw > prev) {
+          console.log('Taking picture for MaxYawR');
+          takePicture('MaxYawR.jpg');
         }
-        return Math.max(prev, roll);
+        return Math.max(prev, yaw);
+      });
+    } else if (yaw < 345 && yaw > 180) {
+      let yawL = 360 - yaw;
+      console.log(`Current YawL: ${yawL}, Max YawL: ${maxYawL}`);
+      setScreenText('<- ' + yawL + '°');
+      setMaxYawL((prev) => {
+        if (yawL > prev) {
+          console.log('Taking picture for MaxYawL');
+          takePicture('MaxYawL.jpg');
+        }
+        return Math.max(prev, yawL);
       });
     } else {
-      //if (roll < 360 && roll > 80) {
-      let rollL = 360 - roll;
-      console.log(`Current RollL: ${rollL}, Max RollL: ${maxRollL}`);
-      setScreenText('<- ' + rollL + '°');
-      setMaxRollL((prev) => {
-        if (rollL > prev) {
-          console.log('Taking picture for MaxRollL');
-          takePicture('MaxRollL.jpg');
-        }
-        return Math.max(prev, rollL);
-      });
-    } /*else {
-            console.log(roll, 'Roll out of range.');
-            setScreenText('');
-        }*/
-  }, [roll]);
+      console.log('Yaw out of range.');
+      setScreenText('');
+    }
+  }, [yaw]);
 
   const handleFacesDetected = ({ faces }) => {
     if (faces.length === 0) {
@@ -109,7 +108,7 @@ const evaluationRoll = () => {
     } else if (faces.length > 0) {
       const face = faces[0];
 
-      setRoll(face.rollAngle.toFixed(0));
+      setYaw(face.yawAngle.toFixed(0));
 
       if (
         face.LEFT_EYE &&
@@ -153,13 +152,12 @@ const evaluationRoll = () => {
 
   const startEvaluation = async () => {
     setCacheBuster(Date.now());
-    clearRollValues();
+    clearYawValues();
     setEvaluationStarted(true);
   };
 
   const startRollEvaluation = () => {
     //  HIER DANN WEITER ZU ROLL derweeil HOME
-    //router.push('tabs');
     router.replace('evaluationRoll');
   };
 
@@ -171,13 +169,12 @@ const evaluationRoll = () => {
       const options = { quality: 1, base64: false };
       const photo = await cameraRef.current.takePictureAsync(options);
       setCapturedPhotoUri(photo.uri);
-      setShowEvaluationResult(true);
     }
   };
 
-  const clearRollValues = () => {
-    setMaxRollR(0);
-    setMaxRollL(0);
+  const clearYawValues = () => {
+    setMaxYawR(0);
+    setMaxYawL(0);
   };
 
   return (
@@ -289,18 +286,18 @@ const evaluationRoll = () => {
         <View style={styles.startScreen}>
           <Image
             source={{
-              uri: `${FileSystem.documentDirectory}MaxRollR.jpg?${cacheBuster}`,
+              uri: `${FileSystem.documentDirectory}MaxYawR.jpg?${cacheBuster}`,
             }}
             style={{ width: 100, height: 100 }}
           />
-          <Text>Max Roll Right: {maxRollR.toFixed(2)}°</Text>
+          <Text>Max Yaw Right: {maxYawR.toFixed(2)}°</Text>
           <Image
             source={{
-              uri: `${FileSystem.documentDirectory}MaxRollL.jpg?${cacheBuster}`,
+              uri: `${FileSystem.documentDirectory}MaxYawL.jpg?${cacheBuster}`,
             }}
             style={{ width: 100, height: 100 }}
           />
-          <Text>Max Roll Left: {maxRollL.toFixed(2)}°</Text>
+          <Text>Max Yaw Left: {maxYawL.toFixed(2)}°</Text>
           <TouchableOpacity onPress={startEvaluation}>
             <Text style={styles.startButton}>
               Evaluation erneut durchführen
@@ -322,13 +319,14 @@ const evaluationRoll = () => {
         <View style={styles.textContainer}>
           <Text style={styles.faceDesc}>{ScreenText}</Text>
           <Text style={styles.maxValues}>
-            Max Right: {maxRollR.toFixed(2)}°, Max Left: {maxRollL.toFixed(2)}°
+            Max Right: {maxYawR.toFixed(2)}°, Max Left: {maxYawL.toFixed(2)}°
           </Text>
         </View>
       )}
     </View>
   );
 };
+
 const styles = StyleSheet.create({
   container: {
     flex: 1,
@@ -375,3 +373,5 @@ const styles = StyleSheet.create({
     color: 'yellow',
   },
 });
+
+export default evaluationYaw;
