+++ app/tabs/barcode.jsx
