+++ app/test.jsx
@@ -1,12 +1,12 @@
 import { View, Text } from 'react-native';
 import React from 'react';
 
-const modal = () => {
+const test = () => {
   return (
     <View>
-      <Text>modal</Text>
+      <Text>test</Text>
     </View>
   );
 };
 
-export default modal;
+export default test;
